Moved some utils around
